# spirv-tools: backport patch to fix clang-10, fix compiler.cppstd validation