--- a/recipes/spirv-tools/all/conandata.yml
+++ b/recipes/spirv-tools/all/conandata.yml
@@ -24,3 +24,16 @@ sources:
   "1.2.198.0":
     url: "https://github.com/KhronosGroup/SPIRV-Tools/archive/refs/tags/sdk-1.2.198.0.tar.gz"
     sha256: "e8a9fec946f8473129374ad6b98ee690ac9a4574ace7cb3b46bbeb4eddfdc33b"
+patches:
+  "1.4.309.0":
+    - patch_file: "patches/0001-fix-clang-20-build-issue.patch"
+      patch_description: "fix clang-20 build issue"
+      patch_type: "backport"
+  "1.3.268.0":
+    - patch_file: "patches/0001-fix-clang-20-build-issue.patch"
+      patch_description: "fix clang-20 build issue"
+      patch_type: "backport"
+  "1.3.261.1":
+    - patch_file: "patches/0001-fix-clang-20-build-issue.patch"
+      patch_description: "fix clang-20 build issue"
+      patch_type: "backport"