@jcar87 2025-04-25
> PR approved upstream

Thanks @elvisdukaj  - please provide a link so that there's traceability. Also CI is failing for some older versions, make sure it's only backported where needed (or where feasible)