@elvisdukaj 2025-04-25
Related PR: https://github.com/KhronosGroup/SPIRV-Tools/pull/6103