@elvisdukaj 2025-04-25
I'll check also the latest two versions and see if the patch is still applicable. Interestingly only macOS fails? Does it mean the different systems are running different versions?