@elvisdukaj 2025-04-26
I built all the listed version and indeed some of them were not applicable.